DBC22-2274: initial refactoring
