+++ src/frontend/src/Components/map/panels/LocalWeatherPanel.scss
@@ -1,0 +1,136 @@
+@import "../../../styles/variables";
+
+.popup {
+  // Local weather popup
+  &--road-weather {
+    .popup__title {
+      background-color: #ECEAE8;
+      border-top: 4px solid $Type-Primary;
+
+      .name {
+        color: $Type-Primary;
+      }
+
+      .sub-name {
+        font-size: 0.875rem;
+        color: $Grey70;
+      }
+
+      &__icon {
+        background-color: $Type-Primary;
+        color: white;
+      }
+    }
+
+    .popup__content {
+      text-align: center;
+      p {
+        margin-bottom: 0;
+      }
+      &__title {
+        margin: 1rem 1rem 0;
+        padding-bottom: 1rem;
+        border-bottom: 1px solid $Divider;
+
+        .name {
+          color: $Type-Primary;
+          font-size: 2.25rem;
+        }
+
+        .friendly-time {
+          margin: 0 auto 1rem;
+          &-text {
+            color: $Type-Secondary;
+            font-size: 0.875rem;
+          }
+        }
+
+        .description {
+          margin-top: 1rem;
+          font-size: 0.875rem;
+        }
+      }
+      &__description {
+        & > div {
+          padding-bottom: 1rem;
+        }
+        .road-condition {
+          .data {
+            font-size: 1.25rem;
+            font-weight: 700;
+            text-transform: lowercase;
+
+            &:first-letter {
+              text-transform: capitalize;
+            }
+
+          }
+        }
+        .temperatures {
+          display: flex;
+          justify-content: space-around;
+
+          .temperature {
+            .data {
+              font-size: 2.25rem;
+              font-weight: 700;
+            }
+          }
+        }
+      }
+    }
+  }
+
+  // Common styles for local weather and Regional weather
+  .data-card {
+    margin: 0.75rem 0;
+    padding: 0.75rem;
+    border-radius: 8px;
+    box-shadow: 0px 6.4px 14.4px 0px rgba(0, 0, 0, 0.13), 0px 1.2px 3.6px 0px rgba(0, 0, 0, 0.10);
+
+    &__row {
+      display: flex;
+
+      & + .data-card__row {
+        margin-top: 0.75rem;
+        padding-top: 0.75rem;
+        border-top: 1px solid $Divider;
+      }
+
+      &:not(.group) {
+        .data-icon {
+          align-self: top;
+        }
+      }
+
+      .data-icon {
+        margin-right: 0.75rem;
+      }
+
+      .data-group {
+        flex: 1 0;
+
+        &__row {
+          display: flex;
+        }
+      }
+    }
+
+    .label, .data {
+      font-size: 0.875rem;
+      margin-bottom: 0;
+    }
+
+    .label {
+      text-align: left;
+      font-weight: 700;
+    }
+
+    .data {
+      text-align: right;
+      margin-left: auto;
+      min-width: 80px;
+      color: $Type-Secondary;
+    }
+  }
+}
