+++ src/frontend/src/Components/map/panels/index.scss
@@ -1,0 +1,119 @@
+@import "../../../styles/variables";
+
+.popup {
+  &__title {
+    background-color: #F1F8FE;
+    border-top: 4px solid #053662;
+    padding: 1rem 1rem 0.5rem;
+
+    .name {
+      margin-top: 0.5rem;
+      margin-bottom: 0;
+      font-size: 1.25rem;
+      color: #053662;
+    }
+
+    &__icon {
+      width: 36px;
+      height: 36px;
+      border-radius: 6px;
+      background-color: #053662;
+      color: white;
+      display: flex;
+      svg {
+        margin: auto;
+      }
+    }
+  }
+
+  &__content {
+    padding-bottom: 1.5rem;
+
+    &__title {
+      margin: 1rem 1rem 0.5rem;
+
+      .name {
+        font-size: 1.25rem;
+        color: #053662;
+      }
+    }
+
+    &__image {
+      position: relative;
+
+      img {
+        width: 100%;
+      }
+
+      .timestamp {
+        background-color: $Black;
+        position: absolute;
+        bottom: 0;
+        padding: 0 10px;
+        display: flex;
+        color: $White;
+        align-items: baseline;
+        width: 100%;
+
+        p {
+          color: $White;
+          margin-bottom: 0;
+          font-size: 0.625rem;
+        }
+
+        .driveBC {
+          font-family: serif;
+          span {
+            color: $BC-Yellow;
+          }
+          margin-right: 10px;
+        }
+
+        .friendly-time, .formatted-date {
+          margin-left: auto;
+        }
+
+        .formatted-date {
+          font-size: 0.625rem;
+        }
+      }
+
+      .camera-unavailable {
+        background-color: $Surface-status-red;
+        padding: 1rem;
+        min-height: 225px;
+        font-size: 1rem;
+        p, .card-img-box {
+          color: $Type-status-red;
+          clear: right;
+        }
+
+        svg {
+          font-size: 1rem;
+        }
+
+        .card-pill {
+          background-color: $Error;
+          border-radius: 12px;
+          height: 24px;
+          float: right;
+          display: flex;
+          justify-content: center;
+          align-items: center;
+          padding: 4px 6px;
+
+          p {
+            color: $White;
+            font-weight: 700;
+            margin-bottom: 0;
+            font-size: 0.625rem;
+          }
+        }
+      }
+    }
+
+    &__description {
+      padding: 1rem 1rem 0;
+    }
+  }
+}
